Move EventFactory to watch module
